# Conflicts:
#	UI/src/SystemUI.java
#	out/production/UI/SystemUI$1.class
#	out/production/UI/SystemUI$OrderTypeOptions.class
#	out/production/UI/SystemUI$StartMenuOptions.class
#	out/production/UI/SystemUI$UpdateProductOptions.class
#	out/production/UI/SystemUI.class

+++ Engine/src/exceptions/UserLocationEqualToStoreException.java
@@ -17,6 +17,6 @@ public class UserLocationEqualToStoreException extends RuntimeException
 
     public String getMessage()
     {
-        return String.format(EXCEPTION_MESSAGE, userLocation.getX(), userLocation.getY());
+        return String.format(EXCEPTION_MESSAGE, (int) userLocation.getX(), (int) userLocation.getY());
     }
 }
